moved logging to cli.py
